+++ IBS/DataAccess/ViewVoucherList.cs
@@ -7,7 +7,7 @@ public partial class ViewVoucherList
 {
     public string? VchrNo { get; set; }
 
-    public byte? Sno { get; set; }
+    public int? Sno { get; set; }
 
     public string? AccDesc { get; set; }
